reorganise the tests folders
