+++ package.json
@@ -77,6 +77,9 @@
   "jest": {
     "preset": "ts-jest",
     "testEnvironment": "node",
+    "setupFilesAfterEnv": [
+      "./src/test/setup.ts"
+    ],
     "globals": {
       "ts-jest": {
         "isolatedModules": true
