+++ src/test/setup.ts
@@ -1,0 +1,17 @@
+import {
+  connect,
+  closeDatabase,
+  clearDatabase,
+} from "../mongo/config_db_testing";
+
+beforeAll(async () => {
+  await connect();
+});
+
+afterEach(async () => {
+  await clearDatabase();
+});
+
+afterAll(async () => {
+  await closeDatabase();
+});
