+++ src/test/unit/server.spec.ts
@@ -1,27 +1,15 @@
 import { createTestClient } from "apollo-server-testing";
-import {
-  connect,
-  closeDatabase,
-  clearDatabase,
-} from "./mongo/config_db_testing";
 
-import { getApolloServer } from "./server";
-import { Task } from "./models";
-import { User } from "./models";
+import { getApolloServer } from "../../server";
+import { User } from "../../models";
 
 describe("Apollo server", () => {
   let mutate: any;
   let query: any;
   beforeEach(async () => {
-    await connect();
     const testClient = createTestClient(await getApolloServer());
     mutate = testClient.mutate;
     query = testClient.query;
-  });
-
-  afterEach(async () => {
-    await clearDatabase();
-    await closeDatabase();
   });
 
   describe("query users", () => {
